[Encointer] fix treasury asset payment (#944)
Even though we have verified that asset payment on Asset Hub works in
general, we have now encointered a protocol bug in the encointer pallets
erroneously registering asset payouts as native payouts. The patch bump
of the encointer-pallets
[fixes](https://github.com/encointer/pallets/pull/434) this.

Additionally, while investigating above issue, we found that there is an
XCM error. While our assets are indeed paid out on asset hub, the
overall XCM result is a `NotHoldingFees` error, which came from the
`ReportError` instruction from the appendix. While this needs some more
investigation, we just removed it, as encointer does not use it for now.
The integration tests have been enhanced to also check successful XCM
execution on Asset Hub.