Fix depencencies (#996)
This PR fixes the dependency mess that we still have in the update
branch.

The issues with encointer's dependencies was that we only introduced
patch bumps for the stable202509 upgrade to keep being aligned with the
polkadot-sdk releases (we messed up the rhythm, as the unstable branch
was a major release for us). The patch bump did not propagate to our
innermost dependencies in the lock file, as it was deemed compatible
with the ones explicitly defined in the root toml.

So this PR fixes (I believe all) of the duplicate substrate dependencies
by running

1.  `cargo update -p encointer-crates-at-wrong-version`
2. `psvm --version polkadot-stable2509-1`