Install solc/resolc via `install-deps` (#1043)
It is better to use the already existing infrastructure instead of
adding another way we to install dependencies.

This pull request also ensures we are forwarding return values in `/cmd`
spawned processes to e.g. fail on errors.

<!-- Remember that you can run `/merge` to enable auto-merge in the PR
-->

<!-- Remember to modify the changelog. If you don't need to modify it,
you can check the following box.
Instead, if you have already modified it, simply delete the following
line. -->

- [x] Does not require a CHANGELOG entry